requirements.txt
openai>=1.0.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0
requests>=2.31.0
typer>=0.9.0
colorama>=0.4.6 